Recode selected(${var},'X') to Stata-friendly version
Added a seried of expressions to pull all selected() expressions and replace them with an expression to test wheter the splot version of that variable (e.g. var_X) is equal to the selected() condition. So, if there is a relevance expression that references a aselect_muliple variable (e.g. selected($(v1},'1'), the new 'if' condition would be v1_1 == 1, referencing the way Stata splits the select_multiple strings.