Improved support for different label languages
Added commands that force the user to keep at most one label language, to pull the language name from the one kept label, and to rename the label to "label" regardless of what language it is. This way, the characteristic "CTO_label" will always be the same, regardless of what language is used.